SpamProvider/python/requirements.txt
Click==7.0
distribute-install-component==0.1
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.1
opencc-python-reimplemented==0.1.4
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
tqdm==4.35.0
Werkzeug==0.15.6
